rag_sec/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
requests==2.32.3
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.22
chromadb==0.5.15
python-dotenv==1.0.1
